6.7.2 - New Morse Signal
New Morse code message library for passive buzzer, with alarm options;
Bug fix in BMP280 and AK8963 libraries, which did not indicate successful reading;
New macro indicating the project name, used to name the created data file and basic Morse output signal;
Adjustments in the main code to incorporate Morse messages indicating malfunction of specific parts

+++ src/lib/AK8963/AK8963.h
@@ -5,7 +5,7 @@
 
 #include "../Magn/Magn.h"
 
-#define MPU9250_ADDRESS_M 0b01101000
+#define MPU9250_ADDRESS_M (uint8_t(0b01101000))
 
 #define MPU9250_MAG_ADDRESS 0b00001100
